refactor: modularize tracker and add tests
