+++ models/__init__.py
@@ -1,0 +1,5 @@
+"""Model definitions for the seesea tracking package."""
+
+from .seqtrack import SeqTrackLSTM
+
+__all__ = ["SeqTrackLSTM"]
