+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import pathlib
+import sys
+
+# Ensure project root on sys.path for module imports
+ROOT = pathlib.Path(__file__).resolve().parents[1]
+if str(ROOT) not in sys.path:
+    sys.path.insert(0, str(ROOT))
