+++ tests/test_models_seqtrack_import.py
@@ -1,0 +1,7 @@
+from models.seqtrack import SeqTrackLSTM
+
+
+def test_seqtracklstm_instantiation() -> None:
+    model = SeqTrackLSTM()
+    assert hasattr(model, "predict")
+    assert hasattr(model, "device")
